refactor: 제미나이 관련 로직 위치 조정 및 함수명 구체화
